Indicadores Dash ok

+++ SistemaVentasBatia/DTOs/DashboardDTO.cs
@@ -5,7 +5,7 @@ namespace SistemaClientesBatia.DTOs
     public class DashboardDTO
     {
         public decimal Asistencia { get; set; }
-        public int Entregas { get; set; }
+        public decimal Entregas { get; set; }
         public int Supervision { get; set; }
         public int Evaluaciones { get; set; }
         public List<AsistenciaMesDTO> AsistenciaMes { get; set; }
